+++ tgPlansBot/eventmanage.go
@@ -33,6 +33,7 @@ func (tgp *TGPlansBot) initEventCommands() {
 	tgp.cmds.Add(tgCommands.Command{Mode: userManager.MODE_EDIT_DATE, Handler: tgp.edit_setDate})
 	tgp.cmds.Add(tgCommands.Command{Mode: userManager.MODE_EDIT_TIME, Handler: tgp.edit_setTime})
 	tgp.cmds.Add(tgCommands.Command{Mode: userManager.MODE_EDIT_PICTURE, Handler: tgp.edit_setPicture, SpecialMode: "photo"})
+	tgp.cmds.Add(tgCommands.Command{Mode: userManager.MODE_EDIT_PUBLIC, Handler: tgp.edit_setPublic, SpecialMode: "public"})
 
 	tgp.cmds.AddCB(tgCommands.Callback{DataPrefix: "calen", Mode: userManager.MODE_CREATE_EVENTDATE, Handler: tgp.create_ClickDate})
 	tgp.cmds.AddCB(tgCommands.Callback{DataPrefix: "time", Mode: userManager.MODE_CREATE_EVENTTIME, Handler: tgp.create_ClickTime})
@@ -73,6 +74,11 @@ func (tgp *TGPlansBot) eventDetails(usrInfo *userManager.UserInfo, chatId int64,
 	if event.Notes() != "" {
 		t += "<b>" + loc.Sprintf("Notes:") + "</b>\n" + event.Notes() + "\n"
 	}
+	isPublic, _, _ := event.Public()
+	if isPublic {
+		t += "\n" + loc.Sprintf("🌎 Event is listed in public directory") + "\n"
+	}
+
 	if event.PictureURL() != "" {
 		t += "\n<i>" + loc.Sprintf("🖼 Event includes a picture") + "</i>\n"
 	}